+++ pkg/policy/l4_filter_test.go
@@ -168,7 +168,7 @@ func TestMergeAllowAllL3AndAllowAllL7(t *testing.T) {
 
 	filter, ok := l4IngressPolicy["80/TCP"]
 	require.True(t, ok)
-	require.Equal(t, 80, filter.Port)
+	require.Equal(t, uint16(80), filter.Port)
 	require.True(t, filter.Ingress)
 
 	require.True(t, filter.SelectsAllEndpoints())
@@ -335,7 +335,7 @@ func TestMergeAllowAllL3AndShadowedL7(t *testing.T) {
 
 	filter, ok := l4IngressPolicy["80/TCP"]
 	require.True(t, ok)
-	require.Equal(t, 80, filter.Port)
+	require.Equal(t, uint16(80), filter.Port)
 	require.True(t, filter.Ingress)
 
 	require.True(t, filter.SelectsAllEndpoints())
